DevEvents/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
djangorestframework==3.15.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
